requirements.txt
# PR Agent Dependencies
# Install with: pip install -r requirements-pr-agent.txt

# Core LangChain components
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20

# LLM Providers (at least one required)
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
openai>=1.12.0
anthropic>=0.18.0

# Search tools (at least one required)
google-search-results>=2.4.2  # Serper API
tavily-python>=0.2.0         # Tavily API

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Already in main requirements (included for completeness)
# pandas>=2.0.0
# numpy>=1.24.0
